trade-service 작동전

+++ marketdata-insight-service/marketdata-insight-application/src/main/java/shop/shportfolio/marketdata/insight/application/command/response/CandleDayTrackResponse.java
@@ -3,19 +3,21 @@ package shop.shportfolio.marketdata.insight.application.command.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class CandleDayTrackResponse {
     private final String market;
     private final String candleDateTimeUtc;
     private final String candleDateTimeKst;
-    private final Double openingPrice;
-    private final Double highPrice;
-    private final Double lowPrice;
-    private final Double tradePrice;
-    private final Double candleAccTradePrice;
-    private final Double candleAccTradeVolume;
-    private final Double prevClosingPrice;
+    private final BigDecimal openingPrice;
+    private final BigDecimal highPrice;
+    private final BigDecimal lowPrice;
+    private final BigDecimal tradePrice;
+    private final BigDecimal candleAccTradePrice;
+    private final BigDecimal candleAccTradeVolume;
+    private final BigDecimal prevClosingPrice;
     private final Double changePrice;
     private final Double changeRate;
 }
