+++ marketdata-insight-service/marketdata-insight-application/src/main/java/shop/shportfolio/marketdata/insight/application/command/response/CandleMinuteTrackResponse.java
@@ -3,17 +3,19 @@ package shop.shportfolio.marketdata.insight.application.command.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class CandleMinuteTrackResponse {
     private final String marketId;
     private final String candleDateTimeKST;
-    private final Double openingPrice;
-    private final Double highPrice;
-    private final Double lowPrice;
-    private final Double tradePrice;
+    private final BigDecimal openingPrice;
+    private final BigDecimal highPrice;
+    private final BigDecimal lowPrice;
+    private final BigDecimal tradePrice;
     private final Long timestamp;
-    private final Double candleAccTradePrice;
-    private final Double candleAccTradeVolume;
+    private final BigDecimal candleAccTradePrice;
+    private final BigDecimal candleAccTradeVolume;
     private final Integer unit;
 }
